+++ packages/ui/src/components/common/TokenSelector.tsx
@@ -5,6 +5,7 @@ import {
   ChainIcon,
   Flex,
   Input,
+  Skeleton,
   Text
 } from '../primitives/index.js'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -13,7 +14,6 @@ import { faChevronLeft } from '@fortawesome/free-solid-svg-icons/faChevronLeft'
 import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons/faMagnifyingGlass'
 import { Modal } from '../common/Modal.js'
 import type { Token } from '../../types/index.js'
-import { ChainTokenIcon } from '../primitives/ChainTokenIcon.js'
 import Fuse from 'fuse.js'
 import ChainFilter, { type ChainFilterValue } from '../common/ChainFilter.js'
 import useRelayClient from '../../hooks/useRelayClient.js'
@@ -264,8 +264,7 @@ const TokenSelector: FC<TokenSelectorProps> = ({
     tokenBalances
   ])
 
-  const isLoading =
-    isLoadingDuneBalances || isLoadingSuggestedTokens || isLoadingTokenList
+  const isLoading = isLoadingSuggestedTokens || isLoadingTokenList
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
   const chainFuse = new Fuse(
@@ -531,6 +530,7 @@ const TokenSelector: FC<TokenSelectorProps> = ({
                         currencyList={list as EnhancedCurrencyList}
                         setCurrencyList={setCurrencyList}
                         selectToken={selectToken}
+                        isLoadingDuneBalances={isLoadingDuneBalances}
                         key={idx}
                       />
                     ) : null
@@ -679,12 +679,14 @@ type CurrencyRowProps = {
   currencyList: EnhancedCurrencyList
   setCurrencyList: (currencyList: EnhancedCurrencyList) => void
   selectToken: (currency: Currency, chainId?: number) => void
+  isLoadingDuneBalances: boolean
 }
 
 const CurrencyRow: FC<CurrencyRowProps> = ({
   currencyList,
   setCurrencyList,
-  selectToken
+  selectToken,
+  isLoadingDuneBalances
 }) => {
   const balance = currencyList.totalBalance
   const decimals =
@@ -761,6 +763,9 @@ const CurrencyRow: FC<CurrencyRowProps> = ({
           </Text>
         ) : null}
       </Flex>
+      {isLoadingDuneBalances && !balance ? (
+        <Skeleton css={{ ml: 'auto', width: 60 }} />
+      ) : null}
       {balance ? (
         <Text color="subtle" style="subtitle3" css={{ ml: 'auto' }}>
           {formatBN(balance, 5, decimals, compactBalance)}